Alterando a Service da entidade Usuario

+++ src/main/java/dev/Zerphyis/library/Service/UsersService.java
@@ -2,7 +2,9 @@ package dev.Zerphyis.library.Service;
 
 import dev.Zerphyis.library.Entity.Datas.DataUsers;
 import dev.Zerphyis.library.Entity.User.Users;
+import dev.Zerphyis.library.Exeception.UserNotFoundExeception;
 import dev.Zerphyis.library.Repository.UsersRepository;
+import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,26 +13,34 @@ import java.util.List;
 @Service
 public class UsersService {
     @Autowired
-    UsersRepository repository;
+    private UsersRepository repository;
 
-    public Users registerUser(DataUsers data){
-        var newUsers= new Users(data);
-        return repository.save(newUsers);
+    public Users registerUser(DataUsers data) {
+        var newUser = new Users(data);
+        return repository.save(newUser);
     }
 
-    public List<Users> getAllUsers(){
+    public List<Users> getAllUsers() {
         return repository.findAll();
     }
 
-    public Users getUserById(Long id){
-        return  repository.findById(id).orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
+    public Users getUserById(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new UserNotFoundExeception("Usuário com ID " + id + " não encontrado"));
     }
 
-    public void deleteUser(Long id){
-         repository.deleteById(id);
+    public void deleteUser(Long id) {
+        if (!repository.existsById(id)) {
+            throw new UserNotFoundExeception("Usuário com ID " + id + " não encontrado");
+        }
+        repository.deleteById(id);
     }
 
-    public Users updateUsers(Long id, DataUsers data){
+    public Users updateUsers(Long id, DataUsers data) {
+        if (!repository.existsById(id)) {
+            throw new UserNotFoundExeception("Usuário com ID " + id + " não encontrado");
+        }
+
         Users user = getUserById(id);
         user.setName(data.name());
         user.setEmail(data.email());
